requirements.txt
fastapi
uvicorn
fpdf2
ollama
python-multipart
transformers
torch
accelerate
librosa
huggingface_hub
numpy
pillow
urllib3<2.0.0
Cython
nemo_toolkit[asr]
